+++ Project_4/makefile
@@ -1,7 +1,2 @@
-all: compile run
-
 compile:
-	g++ -O2 ising.cpp src/lattice.cpp src/mcmc.cpp src/energy_magnetization.cpp src/expectation_values.cpp src/heat_cap_susceptibility.cpp -std=c++11 -I include -o ising.exe -larmadillo
-
-run:
-	./ising.exe
+	g++ -O2 ising.cpp src/lattice.cpp src/mcmc.cpp -I include -std=c++11 -o ising.exe -larmadillo
