+++ Project_4_old/include/energy_magnetization.hpp
