+++ Project_4_old/makefile
@@ -1,0 +1,7 @@
+all: compile run
+
+compile:
+	g++ -O2 ising.cpp src/lattice.cpp src/mcmc.cpp src/energy_magnetization.cpp src/expectation_values.cpp src/heat_cap_susceptibility.cpp -std=c++11 -I include -o ising.exe -larmadillo
+
+run:
+	./ising.exe
