+++ Project_4_old/src/lattice_class.cpp
